Enron Person of Interest Machine Learning Identifier
The goal of this project is to build a person of interest identifier for
the Enron scandal using the database provided. The dataset consists of a
data dictionary with the employee’s name as the Key and then data for
each of the following info points:
'poi', 'salary', 'bonus', 'deferral_payments', 'total_payments',
'loan_advances', 'restricted_stock_deferred', 'deferred_income',
'total_stock_value', 'expenses', 'exercised_stock_options',
'long_term_incentive', 'restricted_stock', 'director_fees',
'shared_receipt_with_poi’
By using the information that we have- especially the hand generated
list of POIs, indicating whether each person is considered a POI or not-
then using the data – to train our algorithm, the final algorithm could
be potentially used on more data to determine people who should be
investigated as they have a high potential of being a person of interest
in the Enron scandal. One can then find other features and values that
can point to the POI.